backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
sqlalchemy>=2.0.28
psycopg2-binary>=2.9.9
langchain>=0.1.13
langchain-community>=0.0.29
litellm>=1.34.0
python-jobspy>=1.1.48
pydantic>=2.6.4
python-multipart>=0.0.9
pypdf>=4.1.0
python-docx>=1.1.0
google-generativeai>=0.8.0
httpx>=0.27.0
playwright>=1.42.0
